players

+++ app/Http/Controllers/Admin/BlogPostController.php
@@ -52,21 +52,22 @@ class BlogPostController extends Controller
      */
     public function update(Request $request)
     {
-        Log::info($request);
         $blogPost = BlogPost::where('id', $request->id)->first();
-        $file_path = public_path() . '/app_images' . '/' . $blogPost->image;
-        unlink($file_path);
-
-        $blogPostImage = $request->file('image');
-        $extension = $blogPostImage->getClientOriginalExtension();
-        $fileName = $blogPostImage->getFilename() . '.' . $extension;
-        Storage::disk('public')->put($fileName,  File::get($blogPostImage));
-
+        if($request->image) {
+            $file_path = public_path() . '/app_images' . '/' . $blogPost->image;
+            if(file_exists($file_path) && !is_dir($file_path)) {
+                unlink($file_path);
+            }
+            unlink($file_path);
+            $blogPostImage = $request->file('image');
+            $extension = $blogPostImage->getClientOriginalExtension();
+            $fileName = $blogPostImage->getFilename() . '.' . $extension;
+            Storage::disk('public')->put($fileName,  File::get($blogPostImage));
+            $blogPost->image = $fileName;
+        }
         $blogPost->title = $request->title;
         $blogPost->description = $request->description;
         $blogPost->short_description = $request->short_description;
-        $blogPost->image = $request->image;
-        $blogPost->image = $fileName;
         if ($blogPost->save()) {
             return redirect()->back()->with(['updateblogPostStatus' => 1]);
         } else {
